backend/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Handling
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
python-multipart==0.0.6

# AI and CrewAI Framework
crewai>=0.121.1
crewai-tools>=0.46.0
openai>=1.3.0,<2.0.0
langchain>=0.1.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-community>=0.0.38,<0.3.0
langchain-core>=0.1.0,<0.3.0
tiktoken>=0.5.0,<1.0.0
tenacity>=8.2.0,<9.0.0

# Search Tools (optional, only if you use Google search API)
google-search-results>=2.4.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0